+++ middleware/isloggedin.js
@@ -1,0 +1,8 @@
+function isLoggedIn(req, res, next){
+	if (req.isAuthenticated()){
+		return next();
+  }
+  res.redirect('/');
+}
+
+module.exports = isLoggedIn
